testdev: demo script
Demo script that uses GUPnP to advertise a simple device spec. The
device has no services, only a presentation URL. Once the presentation
URL is accessed, a redirect to another URL is done.

This handles a simple use case in which a device is advertised by UPnP
daemon, but the daemon does not actually handle the presentation
URL and when accessed performs a redirect to another service (ex. static
web pages, or a web app).